pyproject.toml
[project]
name = "nshutils"
version = "0.19.1"
description = ""
authors = [{ name = "Nima Shoghi", email = "nimashoghi@gmail.com" }]
requires-python = ">=3.10,<4.0"
readme = "README.md"

dependencies = [
    "lazy-loader",
    "numpy",
    "typing-extensions",
    "jaxtyping",
    "beartype",
    "uuid7",
    "nshconfig",
]

[project.optional-dependencies]
snoop = ["pysnooper"]
pprint = ["rich[jupyter]", "treescope"]
extra = ["pysnooper", "rich[jupyter]", "treescope"]

[project.urls]
homepage = "https://github.com/nimashoghi/nshutils"

[tool.poetry.group.dev.dependencies]
basedpyright = "*"
ruff = "*"
ipykernel = "*"
ipywidgets = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.basedpyright]
typeCheckingMode = "standard"
deprecateTypingAliases = true
strictListInference = true
strictDictionaryInference = true
strictSetInference = true
reportPrivateImportUsage = false

[tool.ruff.lint]
select = ["FA102", "FA100"]
ignore = ["F722", "F821", "E731", "E741"]

[tool.ruff.lint.isort]
required-imports = ["from __future__ import annotations"]
